minor refactor and testcases added
